requirements.txt
torch
matplotlib
numpy
torchvision
tqdm
wandb
pyyaml
